server returns a single up/down response based on several checks
...a bit of a WIP, there's much more cleanup to do here
